:recycle: 챌린지 진행 홈 화면 컴포넌트 레이아웃 리팩토링
