+++ Scene/HomeScene/Sources/HomeScene/Views/ChallengeInProgressView.swift
@@ -38,10 +38,22 @@ final class ChallengeInProgressView: UIView {
         return v
     }()
 
+    /// 내 꽃 상위 컴포넌트
+    lazy var myFlowerTopView: MyFlowerTopView = {
+        let v = MyFlowerTopView()
+//        v.delegate = self
+        return v
+    }()
     /// 내 꽃 정보 뷰
     lazy var myFlowerView: MyFlowerView = {
         let v = MyFlowerView()
         v.delegate = self
+        return v
+    }()
+    /// 상대방 꽃 상위 컴포넌트
+    lazy var partnerFlowerTopView: PartnerFlowerTopView = {
+        let v = PartnerFlowerTopView()
+//        v.delegate = self
         return v
     }()
     /// 상대방 꽃 정보 뷰
@@ -87,7 +99,9 @@ final class ChallengeInProgressView: UIView {
         self.addSubviews(self.topChallengeInfoView,
                          self.progressBar,
                          self.nicknameStackView,
+                         self.partnerFlowerTopView,
                          self.partnerFlowerView,
+                         self.myFlowerTopView,
                          self.myFlowerView,
                          self.heartImage,
                          self.nudgeBeeButton,
@@ -113,16 +127,39 @@ final class ChallengeInProgressView: UIView {
             make.trailing.equalToSuperview().inset(24)
         }
         
+        let flowerBottomOffset = UIDevice.current.deviceType == .default ? -10 : 22
+        // --> PartnerFlower
+        self.partnerFlowerTopView.snp.makeConstraints { make in
+            make.top.greaterThanOrEqualTo(self.progressBar.snp.bottom).offset(7)
+            make.leading.equalToSuperview().offset(32)
+            make.width.equalToSuperview().dividedBy(2).offset(-32)
+            make.height.equalToSuperview().multipliedBy(0.12)
+            make.centerX.equalToSuperview().multipliedBy(0.5)
+        }
+        
         self.partnerFlowerView.snp.makeConstraints { make in
-            make.bottom.equalToSuperview().multipliedBy(0.75)
+            make.top.equalTo(self.partnerFlowerTopView.snp.bottom)
             make.centerX.equalToSuperview().multipliedBy(0.5)
             make.width.equalToSuperview().dividedBy(2)
+//            make.height.equalToSuperview().multipliedBy(0.4)
+            make.bottom.equalTo(self.nudgeBeeButton.snp.top).offset(-flowerBottomOffset)
+        }
+        
+        // --> MyFlower
+        self.myFlowerTopView.snp.makeConstraints { make in
+            make.top.greaterThanOrEqualTo(self.progressBar.snp.bottom).offset(7)
+            make.trailing.equalToSuperview().inset(32)
+            make.width.equalToSuperview().dividedBy(2).offset(-32)
+            make.height.equalToSuperview().multipliedBy(0.12)
+            make.centerX.equalToSuperview().multipliedBy(1.5)
         }
         
         self.myFlowerView.snp.makeConstraints { make in
-            make.bottom.equalToSuperview().multipliedBy(0.75)
+            make.top.equalTo(self.myFlowerTopView.snp.bottom)
             make.centerX.equalToSuperview().multipliedBy(1.5)
             make.width.equalToSuperview().dividedBy(2)
+//            make.height.equalToSuperview().multipliedBy(0.4)
+            make.bottom.equalTo(self.nudgeBeeButton.snp.top).offset(-flowerBottomOffset)
         }
         
         self.heartImage.snp.makeConstraints { make in
@@ -130,15 +167,17 @@ final class ChallengeInProgressView: UIView {
             make.centerY.equalTo(self.myFlowerView.snp.centerY)
         }
                 
+        let beeButtonWidthHeight = UIDevice.current.deviceType == .default ? 50 : 57
+//        let nudgeBottomOffset = UIDevice.current.deviceType == .default ? 10 : 38
         self.nudgeBeeButton.snp.makeConstraints { make in
             make.centerX.equalToSuperview()
-            make.width.height.equalTo(57)
+            make.width.height.equalTo(beeButtonWidthHeight)
             make.bottom.equalTo(self.nudgeTitleLabel.snp.top).offset(-8)
         }
         
         self.nudgeTitleLabel.snp.makeConstraints { make in
             make.centerX.equalToSuperview()
-            make.bottom.equalToSuperview().inset(38)
+            make.bottom.equalToSuperview().offset(-38)
         }
         
     }
@@ -150,7 +189,9 @@ final class ChallengeInProgressView: UIView {
                                          myNickname: viewModel.order.myNameText,
                                          partnerNickname: viewModel.order.partenrNameText)
         self.partnerFlowerView.configureInProgress(viewModel: viewModel.partnerFlower)
+        self.partnerFlowerTopView.configureInProgress(viewModel: viewModel.partnerFlower.topViewModel)
         self.myFlowerView.configureInProgress(viewModel: viewModel.myFlower)
+        self.myFlowerTopView.configureInProgress(viewModel: viewModel.myFlower.topViewModel)
         self.heartImage.isHidden = viewModel.isHeartHidden
         self.nudgeTitleLabel.text = viewModel.stickText
     }
